table-browser: move actions and store to flux dir
